backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
aiosqlite
requests
httpx
websocket-client
